+++ packages/core/src/ruleset/validation/__tests__/__fixtures__/aliases/tuple-extended.ts
@@ -1,0 +1,11 @@
+import { RulesetDefinition } from '@stoplight/spectral-core';
+
+import _scope from './scope';
+
+export { ruleset as default };
+
+const ruleset: RulesetDefinition = {
+  extends: [
+    [_scope, 'all'],
+  ],
+};