+++ packages/ruleset-bundler/src/plugins/url.ts
@@ -20,8 +20,8 @@ export const url = ({ fetch }: IO): Plugin => ({
 
     if (importer !== void 0 && isURL(importer)) {
       const url = new URL(importer);
-      // we change https://cdn.skypack.dev/{name-of-the-package}
-      // to https://cdn.skypack.dev/-/lodash@v4.17.21-K6GEbP02mWFnLA45zAmi/dist=es2020,mode=imports/optimized/lodash.js
+      // we change https://esm.sh/{name-of-the-package}
+      // to https://esm.sh/v135/lodash@4.17.21/es2022/lodash.mjs
       if (isAbsolute(id)) {
         url.pathname = id;
       } else {